Add `SSE` estimator for `CatWtdTrj`

+++ causal-hub/src/distributions/categorical.rs
@@ -28,7 +28,7 @@ pub struct CategoricalConditionalProbabilityDistribution {
     parameters: Array2<f64>,
     parameters_size: usize,
     // Fitted statistics.
-    sample_size: Option<usize>,
+    sample_size: Option<f64>,
     sample_log_likelihood: Option<f64>,
 }
 
@@ -297,12 +297,16 @@ impl CatCPD {
 
     /// Returns the sample size of the dataset used to fit the distribution, if any.
     ///
+    /// # Note
+    ///
+    /// The sample size could be non-integer if the distribution was fitted using a weighted dataset.
+    ///
     /// # Returns
     ///
     /// The sample size of the dataset used to fit the distribution.
     ///
     #[inline]
-    pub const fn sample_size(&self) -> Option<usize> {
+    pub const fn sample_size(&self) -> Option<f64> {
         self.sample_size
     }
 
@@ -338,7 +342,7 @@ impl CatCPD {
         state: (L, I),
         conditioning_states: J,
         parameters: Array2<f64>,
-        sample_size: Option<usize>,
+        sample_size: Option<f64>,
         sample_log_likelihood: Option<f64>,
     ) -> Self
     where
